+++ projects/ngx-hipster/src/lib/ngx-error/ngx-error.component.ts
@@ -7,11 +7,12 @@ import { FormControl } from '@angular/forms';
   styleUrls: ['./ngx-error.component.css']
 })
 export class NgxErrorComponent {
-  @Input() label: string;
-  @Input() control: FormControl;
+  @Input() label: string = 'Field';
+  @Input() control: FormControl | null;
 
   isInvalidControl() {
     return (
+      this.control &&
       this.control.invalid &&
       this.control.errors &&
       (this.control.dirty || this.control.touched)